extract sweetspot activity into separate module
