+++ OpenKeychain/src/main/AndroidManifest.xml
@@ -858,7 +858,7 @@
             android:theme="@android:style/Theme.NoDisplay" />
 
         <activity
-            android:name=".ui.ShowNfcSweetspotActivity"
+            android:name="de.cotech.sweetspot.ShowNfcSweetspotActivity"
             android:exported="false"
             android:theme="@style/Theme.Keychain.AppCompat.Transparent.Fullscreen"
             />
