+++ nfcsweetspot/src/main/res/anim/fade_in_quick.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha android:fromAlpha="0.0" android:toAlpha="1.0"
+        android:interpolator="@android:anim/bounce_interpolator"
+        android:duration="400"
+        />
+</set>
