+++ nfcsweetspot/src/main/res/anim/fade_out_quick.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha android:fromAlpha="1.0" android:toAlpha="0.0"
+        android:interpolator="@android:anim/decelerate_interpolator"
+        android:duration="150"
+        />
+</set>
