+++ nfcsweetspot/src/main/res/drawable/ic_circle.xml
